orabug #30084547: improve zval wrapper - initial version with constructors, operator=, to_value getters, array methods, iterators - build fixes for various OSes (integer types conversions)

+++ util/value.cc
@@ -485,7 +485,7 @@ void zvalue::clear()
 void zvalue::reserve(std::size_t size)
 {
 	zval_ptr_dtor(&zv);
-	array_init_size(&zv, size);
+	array_init_size(&zv, static_cast<uint32_t>(size));
 }
 
 void zvalue::reset()
